Added another player and his functions. Reduced lag by a lot

+++ core/src/com/cs110/app/PacketMessage.java
@@ -12,8 +12,8 @@ public class PacketMessage {
 //    public boolean closeAttack = false;
 //    public boolean rangedAttack = false;
 
-    public int xCord;
-    public int yCord;
+    public float xCord;
+    public float yCord;
 
 //    public String message;
